requirements.txt
testinfra
pytest-logging
paramiko
